third_party/blink/web_tests/external/wpt/url/url-constructor.any.worker_include=mailto-expected.txt

This is a testharness.js-based test.
[PASS] Loading data…
[PASS] Parsing: <mailto:/example.com/> against <http://example.org/foo/bar>
[PASS] Parsing: <mailto:example.com/> against <http://example.org/foo/bar>
[PASS] Parsing: <mailto:/example.com/> without base
[PASS] Parsing: <mailto:example.com/> without base
[PASS] Parsing: <#x> against <mailto:x@x.com>
[FAIL] Parsing: <mailto:/../> without base
  assert_equals: href expected "mailto:/" but got "mailto:/../"
[PASS] Parsing: <> without base
[FAIL] Parsing: <mailto://example.com:8080/pathname?search#hash> without base
  assert_equals: href expected "mailto://example.com:8080/pathname?search#hash" but got "mailto://example.com:8080/pathname?search%23hash"
[FAIL] Parsing: <mailto:///test> without base
  assert_equals: pathname expected "/test" but got "///test"
[FAIL] Parsing: <mailto://test/a/../b> without base
  assert_equals: href expected "mailto://test/b" but got "mailto://test/a/../b"
[FAIL] Parsing: <mailto://:443> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <mailto://test:test> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <mailto://[:1]> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
Harness: the test ran to completion.
